#1351 moved peerHandShake to each file
